Fix the FixedHeaders feature
Instead of scrollY, fix the header at the top of the page when the user
scrolls down.

There was a bug with the responsive extension 2.0.0 (the fixed header
was too wide), but version 1.0.7 works fine. So I am including 1.0.7
as a separate http request (in addition to the single script built by
the DataTables wizard).