New build system: allows to use cached JSON
For performance reasons it is important that the table be initially
loaded with a local copy of the Kiva JSON. Later we will add a 'Refresh'
button which will fetch the live JSON using the Kiva API.

In order to further improve perfromance (eliminate an extra http trip),
the JSON will be included in the main.js. To facilitate that, while
still allowing the JSON to be separately updatable (by a cron script),
we now use a simple Makefile-based build system to preprocess main.js
and copy everything into a deployable 'output' directory.

In the future this will also allow us to include and minify all
javascrip/css assets for even better perfromance.

For now, to build the site run 'make all' (or simply 'make') and it will
build the site in 'output'. Run 'make clean' to delete everything in
output. In the future we'll add a deploy target to rsync output to its
remote home.