Switch to m4 as preprocessor
cpp was doing weird stuff with quotes. I guess that's expected, since it
expected C. But it is surprising how few good general-purpose text
preprocessors are available. I tried:

filepp -- it was very slow, and has a perl dependency
gpp -- is obsessed with having a flexible, definabl syntax (why??) to
the point where I needed to change the quote character and it was a
major hassle.

I ended up with m4, which works and is very fast (and way more flexible
than I want). I should probably use one of the gabillion ruby/python
template engines... but at least m4 is a small dependency (and installed
on most workstations already)