Track local javascript dependencies
This commit is the first step in keeping JavaScript dependencies local
(using bower) rather than using the DataTables CDN.

One advantage is that we can pull in only the parts of jquery-ui that we
need. Another is reliability.

Using bower instead of npm because I think npm's nested dependencies are
insane (especially for such a small project).

Manually handling dependencies (they must be added to bower.json,
js/combined.js.in, AND Makefile) is obviously not optimal, but it is
manageable for such a small site. In the future using an auto dependency
finder like Browserify might be worth it.